fixes #265 and removes unnecessary artifacts moved to NServiceBus.Host.AzureCloudService repo
