--- packaging/nuget/nservicebus.hosting.azure.nuspec
@@ -1,29 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-  <metadata xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-    <id>NServiceBus.Hosting.Azure</id>
-    <title>NServiceBus Hosting Azure</title>
-    <version>$version$</version>
-    <authors>NServiceBus Ltd</authors>
-    <owners>NServiceBus Ltd</owners>
-    <licenseUrl>http://particular.net/LicenseAgreement</licenseUrl>
-    <projectUrl>http://particular.net/</projectUrl>
-    <iconUrl>http://s3.amazonaws.com/nuget.images/NServiceBus_32.png</iconUrl>
-    <requireLicenseAcceptance>true</requireLicenseAcceptance>
-    <description>The Azure Cloudservices Host for NServicebus</description>
-    <copyright>Copyright 2010-2014 NServiceBus. All rights reserved</copyright>
-    <tags>nservicebus servicebus msmq cqrs publish subscribe</tags>
-    <dependencies>
-      <dependency id="WindowsAzure.Storage" version="[3.1.0.1, 5.0.0.0)" />
-      <dependency id="Microsoft.Data.OData" version="[5.6.0.0, 6.0.0.0)" />
-      <dependency id="Microsoft.WindowsAzure.ConfigurationManager" version="[2.0.1.0, 3.0.0.0)" />
-      <dependency id="NServiceBus" version="[5.1.2.0, 6.0.0.0)" />
-      <dependency id="System.Spatial" version="[5.6.0.0, 6.0.0.0)" />
-      <dependency id="NServiceBus.Azure" version="[$version$, 7.0.0.0)" />
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="..\..\src\binaries\NServiceBus.Hosting.Azure.???" target="lib\net45" />
-    <file src="..\..\src\lib\Ionic.Zip.dll" target="lib\net45" />
-  </files>
-</package>
