+++ src/NServiceBus.Azure.sln
@@ -1,13 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.30723.0
+VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.Azure", "NServiceBus.Azure\NServiceBus.Azure.csproj", "{12F1D9F1-0A2C-4442-8D18-67DD096C6300}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.Hosting.Azure", "NServiceBus.Hosting.Azure\NServiceBus.Hosting.Azure.csproj", "{6591ED91-F9A1-4CC3-813E-A33E07439D49}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.Hosting.Azure.HostProcess", "NServiceBus.Hosting.Azure.HostProcess\NServiceBus.Hosting.Azure.HostProcess.csproj", "{11B81F23-64C6-4341-94AC-38B3C4C6B1E7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.Azure.Tests", "NServiceBus.Azure.QuickTests\NServiceBus.Azure.Tests.csproj", "{E25B5433-7166-4C10-A76E-C0546F9E2497}"
 EndProject
@@ -18,9 +14,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{642B64
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.Azure.Deployment", "NServiceBus.Azure.Deployment\NServiceBus.Azure.Deployment.csproj", "{82B9D3D2-C2B9-4BC1-AD78-DD4B470DAEE4}"
 	ProjectSection(ProjectDependencies) = postProject
-		{11B81F23-64C6-4341-94AC-38B3C4C6B1E7} = {11B81F23-64C6-4341-94AC-38B3C4C6B1E7}
-		{6591ED91-F9A1-4CC3-813E-A33E07439D49} = {6591ED91-F9A1-4CC3-813E-A33E07439D49}
 		{12F1D9F1-0A2C-4442-8D18-67DD096C6300} = {12F1D9F1-0A2C-4442-8D18-67DD096C6300}
+	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "nuget.packaging", "nuget.packaging", "{FB694847-4392-4F44-B252-BC1FF3E715F0}"
+	ProjectSection(SolutionItems) = preProject
+		..\packaging\nuget\nservicebus.azure.nuspec = ..\packaging\nuget\nservicebus.azure.nuspec
 	EndProjectSection
 EndProject
 Global
@@ -33,14 +32,6 @@ Global
 		{12F1D9F1-0A2C-4442-8D18-67DD096C6300}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{12F1D9F1-0A2C-4442-8D18-67DD096C6300}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{12F1D9F1-0A2C-4442-8D18-67DD096C6300}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6591ED91-F9A1-4CC3-813E-A33E07439D49}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6591ED91-F9A1-4CC3-813E-A33E07439D49}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6591ED91-F9A1-4CC3-813E-A33E07439D49}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6591ED91-F9A1-4CC3-813E-A33E07439D49}.Release|Any CPU.Build.0 = Release|Any CPU
-		{11B81F23-64C6-4341-94AC-38B3C4C6B1E7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{11B81F23-64C6-4341-94AC-38B3C4C6B1E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{11B81F23-64C6-4341-94AC-38B3C4C6B1E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{11B81F23-64C6-4341-94AC-38B3C4C6B1E7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E25B5433-7166-4C10-A76E-C0546F9E2497}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E25B5433-7166-4C10-A76E-C0546F9E2497}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E25B5433-7166-4C10-A76E-C0546F9E2497}.Release|Any CPU.ActiveCfg = Release|Any CPU
