--- src/NServiceBus.Azure/ConfigurationSource/Azure/IAzureConfigurationSettings.cs
@@ -1,8 +1,0 @@
-namespace NServiceBus.Integration.Azure
-{
-    public interface IAzureConfigurationSettings
-    {
-        string GetSetting(string name);
-        bool TryGetSetting(string name, out string setting);
-    }
-}
