--- src/NServiceBus.Azure/ConfigureAzureIntegration.cs
@@ -1,17 +1,0 @@
-namespace NServiceBus
-{
-    using Integration.Azure;
-
-    public static class ConfigureAzureIntegration
-    {
-        public static void AzureConfigurationSource(this BusConfiguration config, string configurationPrefix = null)
-        {
-            var azureConfigSource = new AzureConfigurationSource(new AzureConfigurationSettings())
-            {
-                ConfigurationPrefix = configurationPrefix
-            };
-            
-            config.CustomConfigurationSource(azureConfigSource);
-        }
-    }
-}
