--- src/NServiceBus.Hosting.Azure.HostProcess/Content/Help.txt
@@ -1,54 +1,0 @@
-﻿NServiceBus Message Endpoint Host Service
-
-USAGE:
-	NServiceBus.Hosting.Azure.HostProcess.exe [/install	[/serviceName]
-					[/displayName]
-					[/description]
-					[/endpointConfigurationType]
-					[/instance]
-					[/startManually]
-					[/username]
-					[/password]] |
-			     [/uninstall[/serviceName]
-					[/instance]]
-										
-OPTIONS:
-
-/install	Install the message endpoint as a Windows service
-/serviceName	Specify the service name for the installed service
-/displayName	Friendly name for the installed service
-/description	Description for the service
-/endpointConfigurationType	 Specify the type implementing IConfigureThisEndpoint that should be used
-/instance	Specify the unique name of the service instance you wish to install
-		ex: "NServiceBus.Host.exe /install /instance:Instance5"
-/startManually	Specifies that the service should start manually
-/username	Username for the account the service should run under
-/password	Password for the service account
-
-If no service name is specified NServiceBus will use the full name of the 
-endpoint configuration type (that which implements NServiceBus.IConfigureThisEndpoint)
-along with the version number of the assembly it is contained within, for example:
-
-	MyPublisher.Endpoint_v1.0.0.0
-	
-The default for the display name is the same value as the service name, and the description
-defaults to a generic NServiceBus host description.
-
-You can also specify the endpoint configuration type in the file NServiceBus.Host.exe.config.
-This file is optional.
-
-If you don't specify the endpoint configuration type either in the command-line or in the
-NServiceBus.Host.exe.config file, all the DLLs in the runtime directory will be scanned
-for a type that implements NServiceBus.IConfigureThisEndpoint.
-
-If you set the service name and/or instance name during installation you will need to specify
-them when uninstalling them as well, ex:
-
-	NServiceBus.Host.exe /uninstall /serviceName:"MyPublisher" /instance:Instance89
-	
-EXAMPLES:
-	NServiceBus.Host.exe /install /serviceName:"MyPublisher" /displayName:"My Publisher Service"
-		/description:"Service for publishing event messages" 
-		/endpointConfigurationType:"YourEndpointConfigType.YourNameSpace, YourAssembly"
-		/username:"corp\serviceuser"
-		/password:"p@ssw0rd!" NServiceBus.Production
