--- src/NServiceBus.Hosting.Azure.HostProcess/Properties/AssemblyInfo.cs
@@ -1,11 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("NServiceBus.Hosting.Azure.HostProcess")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyCopyright("Copyright 2010-2013 NServiceBus. All rights reserved")]
-[assembly: AssemblyProduct("NServiceBus.Hosting.Azure.HostProcess")]
-[assembly: AssemblyCompany("NServiceBus Ltd.")]
-[assembly: AssemblyConfiguration("release")]
-[assembly: ComVisible(false)]
-
