--- src/NServiceBus.Hosting.Azure.HostProcess/app.config
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/></startup></configuration>
