--- src/NServiceBus.Hosting.Azure.HostProcess/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="CommonServiceLocator" version="1.0" targetFramework="net45" />
-  <package id="GitVersionTask" version="1.3.3" targetFramework="net45" developmentDependency="true" />
-  <package id="log4net" version="1.2.10.0" targetFramework="net45" />
-  <package id="NServiceBus" version="5.2.0" targetFramework="net45" />
-  <package id="NuGetPackager" version="0.4.13" targetFramework="net45" developmentDependency="true" />
-</packages>
