--- src/NServiceBus.Hosting.Azure/DynamicHost/EndpointsEventArgs.cs
@@ -1,10 +1,0 @@
-namespace NServiceBus.Hosting
-{
-    using System;
-    using System.Collections.Generic;
-
-    class EndpointsEventArgs : EventArgs
-    {
-        public IEnumerable<EndpointToHost> Endpoints { get; set; }
-    }
-}
