--- src/NServiceBus.Hosting.Azure/FodyWeavers.xml
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Weavers>
-  <Obsolete />
-</Weavers>
