--- src/NServiceBus.Hosting.Azure/Profiles/Development.cs
@@ -1,9 +1,0 @@
-namespace NServiceBus
-{
-    /// <summary>
-    /// Indicates that the infrastructure should configure itself for development.
-    /// </summary>
-    public class Development : IProfile
-    {
-    }
-}
