--- src/NServiceBus.Hosting.Azure/Properties/AssemblyInfo.cs
@@ -1,14 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("NServiceBus.Hosting.Azure")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyCopyright("Copyright 2010-2013 NServiceBus. All rights reserved")]
-[assembly: AssemblyProduct("NServiceBus.Hosting.Azure")]
-[assembly: AssemblyCompany("NServiceBus Ltd.")]
-[assembly: AssemblyConfiguration("release")]
-[assembly: ComVisible(false)]
-[assembly: InternalsVisibleTo("NServiceBus.Azure.Tests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100dde965e6172e019ac82c2639ffe494dd2e7dd16347c34762a05732b492e110f2e4e2e1b5ef2d85c848ccfb671ee20a47c8d1376276708dc30a90ff1121b647ba3b7259a6bc383b2034938ef0e275b58b920375ac605076178123693c6c4f1331661a62eba28c249386855637780e3ff5f23a6d854700eaa6803ef48907513b92")]		
-
-
