--- src/NServiceBus.Hosting.Azure/Roles/AsA_Host.cs
@@ -1,7 +1,0 @@
-namespace NServiceBus
-{
-    /// <summary>
-    /// Indicates this endpoint is a host that merely loads other processes.
-    /// </summary>
-    public interface AsA_Host { }
-}
