--- src/NServiceBus.Hosting.Azure/Roles/AsA_Worker.cs
@@ -1,7 +1,0 @@
-namespace NServiceBus
-{
-    /// <summary>
-    /// Indicates this endpoint is a server.
-    /// </summary>
-    public interface AsA_Worker {}
-}
