--- src/NServiceBus.Hosting.Azure/packages.config
@@ -1,15 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Fody" version="1.25.0" targetFramework="net45" developmentDependency="true" />
-  <package id="GitVersionTask" version="1.3.3" targetFramework="net45" developmentDependency="true" />
-  <package id="Microsoft.Data.Edm" version="5.6.2" targetFramework="net45" />
-  <package id="Microsoft.Data.OData" version="5.6.2" targetFramework="net45" />
-  <package id="Microsoft.Data.Services.Client" version="5.6.2" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="2.0.3" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="5.0.8" targetFramework="net45" />
-  <package id="NServiceBus" version="5.2.0" targetFramework="net45" />
-  <package id="NuGetPackager" version="0.4.13" targetFramework="net45" developmentDependency="true" />
-  <package id="Obsolete.Fody" version="3.1.0.0" targetFramework="net45" developmentDependency="true" />
-  <package id="System.Spatial" version="5.6.2" targetFramework="net45" />
-  <package id="WindowsAzure.Storage" version="4.3.0" targetFramework="net45" />
-</packages>
